+++ src/main/java/subway/application/LineService.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import subway.dao.LineDao;
-import subway.domain.Line;
+import subway.domain.core.Line;
 import subway.dto.request.LineCreateRequest;
 import subway.dto.response.LineResponse;
