+++ src/main/java/subway/application/SectionService.java
@@ -9,9 +9,9 @@ import org.springframework.transaction.annotation.Transactional;
 import subway.dao.LineDao;
 import subway.dao.SectionDao;
 import subway.dao.StationDao;
-import subway.domain.Section;
-import subway.domain.Sections;
-import subway.domain.Station;
+import subway.domain.core.Section;
+import subway.domain.core.Sections;
+import subway.domain.core.Station;
 import subway.dto.request.SectionDeleteRequest;
 import subway.dto.request.SectionUpdateRequest;
 import subway.dto.response.SectionResponse;
