+++ src/main/java/subway/dao/LineDao.java
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
-import subway.domain.Line;
+import subway.domain.core.Line;
 
 @Repository
 public class LineDao {
