+++ src/main/java/subway/dao/StationDao.java
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
-import subway.domain.Station;
+import subway.domain.core.Station;
 
 @Repository
 public class StationDao {
