+++ src/main/java/subway/domain/core/Line.java
@@ -1,20 +1,17 @@
-package subway.domain;
+package subway.domain.core;
 
 import java.util.Objects;
 
-public class Station {
+public class Line {
 	private Long id;
 	private String name;
 
-	public Station() {
-	}
-
-	public Station(Long id, String name) {
-		this.id = id;
+	public Line(final String name) {
 		this.name = name;
 	}
 
-	public Station(String name) {
+	public Line(final Long id, final String name) {
+		this.id = id;
 		this.name = name;
 	}
 
@@ -32,21 +29,12 @@ public class Station {
 			return true;
 		if (o == null || getClass() != o.getClass())
 			return false;
-
-		Station station = (Station)o;
-
-		return Objects.equals(name, station.name);
+		final Line line = (Line)o;
+		return Objects.equals(name, line.name);
 	}
 
 	@Override
 	public int hashCode() {
-		return name != null ? name.hashCode() : 0;
-	}
-
-	@Override
-	public String toString() {
-		return "Station{" +
-			"name='" + name + '\'' +
-			'}';
+		return Objects.hash(name);
 	}
 }
