+++ src/main/java/subway/domain/core/Section.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.core;
 
 public class Section {
 	private Long id;
